Network.AWS.WAF.CreateByteMatchSet

Description

Creates a ByteMatchSet . You then use UpdateByteMatchSet to identify the part of a web request that you want AWS WAF to inspect, such as the values of the User-Agent header or the query string. For example, you can create a ByteMatchSet that matches any requests with User-Agent headers that contain the string BadBot . You can then configure AWS WAF to reject those requests.

To create and configure a ByteMatchSet , perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateByteMatchSet request.

Submit a CreateByteMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateByteMatchSet request.

Submit an UpdateByteMatchSet request to specify the part of the request that you want AWS WAF to inspect (for example, the header or the URI) and the value that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

createByteMatchSet

Arguments

:: Text
cbmsName

-> Text
cbmsChangeToken

-> CreateByteMatchSet

Creates a value of CreateByteMatchSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbmsName - A friendly name or description of the ByteMatchSet . You can't change Name after you create a ByteMatchSet .

cbmsChangeToken - The value returned by the most recent call to GetChangeToken .

data CreateByteMatchSet

See: createByteMatchSet smart constructor.

Request Lenses

cbmsName :: Lens' CreateByteMatchSet Text

A friendly name or description of the ByteMatchSet . You can't change Name after you create a ByteMatchSet .

cbmsChangeToken :: Lens' CreateByteMatchSet Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

createByteMatchSetResponse

Arguments

:: Int
cbmsrsResponseStatus

-> CreateByteMatchSetResponse

Creates a value of CreateByteMatchSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbmsrsByteMatchSet - A ByteMatchSet that contains no ByteMatchTuple objects.

cbmsrsChangeToken - The ChangeToken that you used to submit the CreateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

cbmsrsResponseStatus - -- | The response status code.

data CreateByteMatchSetResponse

See: createByteMatchSetResponse smart constructor.

Response Lenses

cbmsrsByteMatchSet :: Lens' CreateByteMatchSetResponse (Maybe ByteMatchSet)

A ByteMatchSet that contains no ByteMatchTuple objects.

cbmsrsChangeToken :: Lens' CreateByteMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateByteMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

cbmsrsResponseStatus :: Lens' CreateByteMatchSetResponse Int

- | The response status code.