Network.AWS.WAF.CreateSqlInjectionMatchSet

Description

Creates a SqlInjectionMatchSet , which you use to allow, block, or count requests that contain snippets of SQL code in a specified part of web requests. AWS WAF searches for character sequences that are likely to be malicious strings.

To create and configure a SqlInjectionMatchSet , perform the following steps:

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a CreateSqlInjectionMatchSet request.

Submit a CreateSqlInjectionMatchSet request.

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateSqlInjectionMatchSet request.

Submit an UpdateSqlInjectionMatchSet request to specify the parts of web requests in which you want to allow, block, or count malicious SQL code.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

createSqlInjectionMatchSet

Arguments

:: Text
csimsName

-> Text
csimsChangeToken

-> CreateSqlInjectionMatchSet

Creates a value of CreateSqlInjectionMatchSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csimsName - A friendly name or description for the SqlInjectionMatchSet that you're creating. You can't change Name after you create the SqlInjectionMatchSet .

csimsChangeToken - The value returned by the most recent call to GetChangeToken .

data CreateSqlInjectionMatchSet

A request to create a SqlInjectionMatchSet .

See: createSqlInjectionMatchSet smart constructor.

Request Lenses

csimsName :: Lens' CreateSqlInjectionMatchSet Text

A friendly name or description for the SqlInjectionMatchSet that you're creating. You can't change Name after you create the SqlInjectionMatchSet .

csimsChangeToken :: Lens' CreateSqlInjectionMatchSet Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

createSqlInjectionMatchSetResponse

Arguments

:: Int
csimsrsResponseStatus

-> CreateSqlInjectionMatchSetResponse

Creates a value of CreateSqlInjectionMatchSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csimsrsSqlInjectionMatchSet - A SqlInjectionMatchSet .

csimsrsChangeToken - The ChangeToken that you used to submit the CreateSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

csimsrsResponseStatus - -- | The response status code.

data CreateSqlInjectionMatchSetResponse

The response to a CreateSqlInjectionMatchSet request.

See: createSqlInjectionMatchSetResponse smart constructor.

Response Lenses

csimsrsSqlInjectionMatchSet :: Lens' CreateSqlInjectionMatchSetResponse (Maybe SqlInjectionMatchSet)

A SqlInjectionMatchSet .

csimsrsChangeToken :: Lens' CreateSqlInjectionMatchSetResponse (Maybe Text)

The ChangeToken that you used to submit the CreateSqlInjectionMatchSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

csimsrsResponseStatus :: Lens' CreateSqlInjectionMatchSetResponse Int

- | The response status code.