Network.AWS.WAF.DeleteRateBasedRule

Description

Permanently deletes a RateBasedRule . You can't delete a rule if it's still used in any WebACL objects or if it still includes any predicates, such as ByteMatchSet objects.

If you just want to remove a rule from a WebACL , use UpdateWebACL .

To permanently delete a RateBasedRule from AWS WAF, perform the following steps:

Update the RateBasedRule to remove predicates, if any. For more information, see UpdateRateBasedRule .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteRateBasedRule request.

Submit a DeleteRateBasedRule request.

Creating a Request

deleteRateBasedRule

Arguments

:: Text
drbrRuleId

-> Text
drbrChangeToken

-> DeleteRateBasedRule

Creates a value of DeleteRateBasedRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drbrRuleId - The RuleId of the RateBasedRule that you want to delete. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules .

drbrChangeToken - The value returned by the most recent call to GetChangeToken .

data DeleteRateBasedRule

See: deleteRateBasedRule smart constructor.

Request Lenses

drbrRuleId :: Lens' DeleteRateBasedRule Text

The RuleId of the RateBasedRule that you want to delete. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules .

drbrChangeToken :: Lens' DeleteRateBasedRule Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

deleteRateBasedRuleResponse

Arguments

:: Int
drbrrsResponseStatus

-> DeleteRateBasedRuleResponse

Creates a value of DeleteRateBasedRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drbrrsChangeToken - The ChangeToken that you used to submit the DeleteRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

drbrrsResponseStatus - -- | The response status code.

data DeleteRateBasedRuleResponse

See: deleteRateBasedRuleResponse smart constructor.

Response Lenses

drbrrsChangeToken :: Lens' DeleteRateBasedRuleResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

drbrrsResponseStatus :: Lens' DeleteRateBasedRuleResponse Int

- | The response status code.