Network.AWS.WAF.DeleteWebACL

Description

Permanently deletes a WebACL . You can't delete a WebACL if it still contains any Rules .

To delete a WebACL , perform the following steps:

Update the WebACL to remove Rules , if any. For more information, see UpdateWebACL .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of a DeleteWebACL request.

Submit a DeleteWebACL request.

Creating a Request

deleteWebACL

Arguments

:: Text
dwaWebACLId

-> Text
dwaChangeToken

-> DeleteWebACL

Creates a value of DeleteWebACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwaWebACLId - The WebACLId of the WebACL that you want to delete. WebACLId is returned by CreateWebACL and by ListWebACLs .

dwaChangeToken - The value returned by the most recent call to GetChangeToken .

data DeleteWebACL

See: deleteWebACL smart constructor.

Request Lenses

dwaWebACLId :: Lens' DeleteWebACL Text

The WebACLId of the WebACL that you want to delete. WebACLId is returned by CreateWebACL and by ListWebACLs .

dwaChangeToken :: Lens' DeleteWebACL Text

The value returned by the most recent call to GetChangeToken .

Destructuring the Response

deleteWebACLResponse

Arguments

:: Int
dwarsResponseStatus

-> DeleteWebACLResponse

Creates a value of DeleteWebACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dwarsChangeToken - The ChangeToken that you used to submit the DeleteWebACL request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

dwarsResponseStatus - -- | The response status code.

data DeleteWebACLResponse

See: deleteWebACLResponse smart constructor.

Response Lenses

dwarsChangeToken :: Lens' DeleteWebACLResponse (Maybe Text)

The ChangeToken that you used to submit the DeleteWebACL request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

dwarsResponseStatus :: Lens' DeleteWebACLResponse Int

- | The response status code.