Network.AWS.WAF.ListRules

Description

Returns an array of RuleSummary objects.

This operation returns paginated results.

Creating a Request

listRules :: ListRules

Creates a value of ListRules with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrNextMarker - If you specify a value for Limit and you have more Rules than the value of Limit , AWS WAF returns a NextMarker value in the response that allows you to list another group of Rules . For the second and subsequent ListRules requests, specify the value of NextMarker from the previous response to get information about another batch of Rules .

lrLimit - Specifies the number of Rules that you want AWS WAF to return for this request. If you have more Rules than the number that you specify for Limit , the response includes a NextMarker value that you can use to get another batch of Rules .

data ListRules

See: listRules smart constructor.

Request Lenses

lrNextMarker :: Lens' ListRules (Maybe Text)

If you specify a value for Limit and you have more Rules than the value of Limit , AWS WAF returns a NextMarker value in the response that allows you to list another group of Rules . For the second and subsequent ListRules requests, specify the value of NextMarker from the previous response to get information about another batch of Rules .

lrLimit :: Lens' ListRules (Maybe Natural)

Specifies the number of Rules that you want AWS WAF to return for this request. If you have more Rules than the number that you specify for Limit , the response includes a NextMarker value that you can use to get another batch of Rules .

Destructuring the Response

listRulesResponse

Arguments

:: Int
lrrsResponseStatus

-> ListRulesResponse

Creates a value of ListRulesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrrsRules - An array of RuleSummary objects.

lrrsNextMarker - If you have more Rules than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more Rules , submit another ListRules request, and specify the NextMarker value from the response in the NextMarker value in the next request.

lrrsResponseStatus - -- | The response status code.

data ListRulesResponse

See: listRulesResponse smart constructor.

Response Lenses

lrrsRules :: Lens' ListRulesResponse [RuleSummary]

An array of RuleSummary objects.

lrrsNextMarker :: Lens' ListRulesResponse (Maybe Text)

If you have more Rules than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more Rules , submit another ListRules request, and specify the NextMarker value from the response in the NextMarker value in the next request.

lrrsResponseStatus :: Lens' ListRulesResponse Int

- | The response status code.