Network.AWS.WAF.ListSqlInjectionMatchSets

Description

Returns an array of SqlInjectionMatchSet objects.

This operation returns paginated results.

Creating a Request

listSqlInjectionMatchSets :: ListSqlInjectionMatchSets

Creates a value of ListSqlInjectionMatchSets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsimsNextMarker - If you specify a value for Limit and you have more SqlInjectionMatchSet objects than the value of Limit , AWS WAF returns a NextMarker value in the response that allows you to list another group of SqlInjectionMatchSets . For the second and subsequent ListSqlInjectionMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of SqlInjectionMatchSets .

lsimsLimit - Specifies the number of SqlInjectionMatchSet objects that you want AWS WAF to return for this request. If you have more SqlInjectionMatchSet objects than the number you specify for Limit , the response includes a NextMarker value that you can use to get another batch of Rules .

data ListSqlInjectionMatchSets

A request to list the SqlInjectionMatchSet objects created by the current AWS account.

See: listSqlInjectionMatchSets smart constructor.

Request Lenses

lsimsNextMarker :: Lens' ListSqlInjectionMatchSets (Maybe Text)

If you specify a value for Limit and you have more SqlInjectionMatchSet objects than the value of Limit , AWS WAF returns a NextMarker value in the response that allows you to list another group of SqlInjectionMatchSets . For the second and subsequent ListSqlInjectionMatchSets requests, specify the value of NextMarker from the previous response to get information about another batch of SqlInjectionMatchSets .

lsimsLimit :: Lens' ListSqlInjectionMatchSets (Maybe Natural)

Specifies the number of SqlInjectionMatchSet objects that you want AWS WAF to return for this request. If you have more SqlInjectionMatchSet objects than the number you specify for Limit , the response includes a NextMarker value that you can use to get another batch of Rules .

Destructuring the Response

listSqlInjectionMatchSetsResponse

Arguments

:: Int
lsimsrsResponseStatus

-> ListSqlInjectionMatchSetsResponse

Creates a value of ListSqlInjectionMatchSetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lsimsrsNextMarker - If you have more SqlInjectionMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more SqlInjectionMatchSet objects, submit another ListSqlInjectionMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

lsimsrsSqlInjectionMatchSets - An array of SqlInjectionMatchSetSummary objects.

lsimsrsResponseStatus - -- | The response status code.

data ListSqlInjectionMatchSetsResponse

The response to a ListSqlInjectionMatchSets request.

See: listSqlInjectionMatchSetsResponse smart constructor.

Response Lenses

lsimsrsNextMarker :: Lens' ListSqlInjectionMatchSetsResponse (Maybe Text)

If you have more SqlInjectionMatchSet objects than the number that you specified for Limit in the request, the response includes a NextMarker value. To list more SqlInjectionMatchSet objects, submit another ListSqlInjectionMatchSets request, and specify the NextMarker value from the response in the NextMarker value in the next request.

lsimsrsSqlInjectionMatchSets :: Lens' ListSqlInjectionMatchSetsResponse [SqlInjectionMatchSetSummary]

An array of SqlInjectionMatchSetSummary objects.

lsimsrsResponseStatus :: Lens' ListSqlInjectionMatchSetsResponse Int

- | The response status code.