Network.AWS.WAF.UpdateRateBasedRule

Description

Inserts or deletes Predicate objects in a rule and updates the RateLimit in the rule.

Each Predicate object identifies a predicate, such as a ByteMatchSet or an IPSet , that specifies the web requests that you want to block or count. The RateLimit specifies the number of requests every five minutes that triggers the rule.

If you add more than one predicate to a RateBasedRule , a request must match all the predicates and exceed the RateLimit to be counted or blocked. For example, suppose you add the following to a RateBasedRule :

An IPSet that matches the IP address 192.0.2.44/32

A ByteMatchSet that matches BadBot in the User-Agent header

Further, you specify a RateLimit of 15,000.

You then add the RateBasedRule to a WebACL and specify that you want to block requests that satisfy the rule. For a request to be blocked, it must come from the IP address 192.0.2.44 and the User-Agent header in the request must contain the value BadBot . Further, requests that match these two conditions much be received at a rate of more than 15,000 every five minutes. If the rate drops below this limit, AWS WAF no longer blocks the requests.

As a second example, suppose you want to limit requests to a particular page on your site. To do this, you could add the following to a RateBasedRule :

A ByteMatchSet with FieldToMatch of URI

A PositionalConstraint of STARTS_WITH

A TargetString of login

Further, you specify a RateLimit of 15,000.

By adding this RateBasedRule to a WebACL , you could limit requests to your login page without affecting the rest of your site.

Creating a Request

updateRateBasedRule

Arguments

:: Text
urbrRuleId

-> Text
urbrChangeToken

-> Natural
urbrRateLimit

-> UpdateRateBasedRule

Creates a value of UpdateRateBasedRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urbrRuleId - The RuleId of the RateBasedRule that you want to update. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules .

urbrChangeToken - The value returned by the most recent call to GetChangeToken .

urbrUpdates - An array of RuleUpdate objects that you want to insert into or delete from a RateBasedRule .

urbrRateLimit - The maximum number of requests, which have an identical value in the field specified by the RateKey , allowed in a five-minute period. If the number of requests exceeds the RateLimit and the other predicates specified in the rule are also met, AWS WAF triggers the action that is specified for this rule.

data UpdateRateBasedRule

See: updateRateBasedRule smart constructor.

Request Lenses

urbrRuleId :: Lens' UpdateRateBasedRule Text

The RuleId of the RateBasedRule that you want to update. RuleId is returned by CreateRateBasedRule and by ListRateBasedRules .

urbrChangeToken :: Lens' UpdateRateBasedRule Text

The value returned by the most recent call to GetChangeToken .

urbrUpdates :: Lens' UpdateRateBasedRule [RuleUpdate]

An array of RuleUpdate objects that you want to insert into or delete from a RateBasedRule .

urbrRateLimit :: Lens' UpdateRateBasedRule Natural

The maximum number of requests, which have an identical value in the field specified by the RateKey , allowed in a five-minute period. If the number of requests exceeds the RateLimit and the other predicates specified in the rule are also met, AWS WAF triggers the action that is specified for this rule.

Destructuring the Response

updateRateBasedRuleResponse

Arguments

:: Int
urbrrsResponseStatus

-> UpdateRateBasedRuleResponse

Creates a value of UpdateRateBasedRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urbrrsChangeToken - The ChangeToken that you used to submit the UpdateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

urbrrsResponseStatus - -- | The response status code.

data UpdateRateBasedRuleResponse

See: updateRateBasedRuleResponse smart constructor.

Response Lenses

urbrrsChangeToken :: Lens' UpdateRateBasedRuleResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateRateBasedRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

urbrrsResponseStatus :: Lens' UpdateRateBasedRuleResponse Int

- | The response status code.