Network.AWS.WAF.UpdateRule

Description

Inserts or deletes Predicate objects in a Rule . Each Predicate object identifies a predicate, such as a ByteMatchSet or an IPSet , that specifies the web requests that you want to allow, block, or count. If you add more than one predicate to a Rule , a request must match all of the specifications to be allowed, blocked, or counted. For example, suppose you add the following to a Rule :

A ByteMatchSet that matches the value BadBot in the User-Agent header

An IPSet that matches the IP address 192.0.2.44

You then add the Rule to a WebACL and specify that you want to block requests that satisfy the Rule . For a request to be blocked, the User-Agent header in the request must contain the value BadBot and the request must originate from the IP address 192.0.2.44.

To create and configure a Rule , perform the following steps:

Create and update the predicates that you want to include in the Rule .

Create the Rule . See CreateRule .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateRule request.

Submit an UpdateRule request to add predicates to the Rule .

Create and update a WebACL that contains the Rule . See CreateWebACL .

If you want to replace one ByteMatchSet or IPSet with another, you delete the existing one and add the new one.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

updateRule

Arguments

:: Text
urRuleId

-> Text
urChangeToken

-> UpdateRule

Creates a value of UpdateRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urRuleId - The RuleId of the Rule that you want to update. RuleId is returned by CreateRule and by ListRules .

urChangeToken - The value returned by the most recent call to GetChangeToken .

urUpdates - An array of RuleUpdate objects that you want to insert into or delete from a Rule . For more information, see the applicable data types: * RuleUpdate : Contains Action and Predicate * Predicate : Contains DataId , Negated , and Type * FieldToMatch : Contains Data and Type

data UpdateRule

See: updateRule smart constructor.

Request Lenses

urRuleId :: Lens' UpdateRule Text

The RuleId of the Rule that you want to update. RuleId is returned by CreateRule and by ListRules .

urChangeToken :: Lens' UpdateRule Text

The value returned by the most recent call to GetChangeToken .

urUpdates :: Lens' UpdateRule [RuleUpdate]

An array of RuleUpdate objects that you want to insert into or delete from a Rule . For more information, see the applicable data types: * RuleUpdate : Contains Action and Predicate * Predicate : Contains DataId , Negated , and Type * FieldToMatch : Contains Data and Type

Destructuring the Response

updateRuleResponse

Arguments

:: Int
urrsResponseStatus

-> UpdateRuleResponse

Creates a value of UpdateRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urrsChangeToken - The ChangeToken that you used to submit the UpdateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

urrsResponseStatus - -- | The response status code.

data UpdateRuleResponse

See: updateRuleResponse smart constructor.

Response Lenses

urrsChangeToken :: Lens' UpdateRuleResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateRule request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

urrsResponseStatus :: Lens' UpdateRuleResponse Int

- | The response status code.