Network.AWS.WAF.UpdateSizeConstraintSet

Description

Inserts or deletes SizeConstraint objects (filters) in a SizeConstraintSet . For each SizeConstraint object, you specify the following values:

Whether to insert or delete the object from the array. If you want to change a SizeConstraintSetUpdate object, you delete the existing object and add a new one.

The part of a web request that you want AWS WAF to evaluate, such as the length of a query string or the length of the User-Agent header.

Whether to perform any transformations on the request, such as converting it to lowercase, before checking its length. Note that transformations of the request body are not supported because the AWS resource forwards only the first 8192 bytes of your request to AWS WAF.

A ComparisonOperator used for evaluating the selected part of the request against the specified Size , such as equals, greater than, less than, and so on.

The length, in bytes, that you want AWS WAF to watch for in selected part of the request. The length is computed after applying the transformation.

For example, you can add a SizeConstraintSetUpdate object that matches web requests in which the length of the User-Agent header is greater than 100 bytes. You can then configure AWS WAF to block those requests.

To create and configure a SizeConstraintSet , perform the following steps:

Create a SizeConstraintSet. For more information, see CreateSizeConstraintSet .

Use GetChangeToken to get the change token that you provide in the ChangeToken parameter of an UpdateSizeConstraintSet request.

Submit an UpdateSizeConstraintSet request to specify the part of the request that you want AWS WAF to inspect (for example, the header or the URI) and the value that you want AWS WAF to watch for.

For more information about how to use the AWS WAF API to allow or block HTTP requests, see the AWS WAF Developer Guide .

Creating a Request

updateSizeConstraintSet

Arguments

:: Text
uscsSizeConstraintSetId

-> Text
uscsChangeToken

-> UpdateSizeConstraintSet

Creates a value of UpdateSizeConstraintSet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uscsSizeConstraintSetId - The SizeConstraintSetId of the SizeConstraintSet that you want to update. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets .

uscsChangeToken - The value returned by the most recent call to GetChangeToken .

uscsUpdates - An array of SizeConstraintSetUpdate objects that you want to insert into or delete from a SizeConstraintSet . For more information, see the applicable data types: * SizeConstraintSetUpdate : Contains Action and SizeConstraint * SizeConstraint : Contains FieldToMatch , TextTransformation , ComparisonOperator , and Size * FieldToMatch : Contains Data and Type

data UpdateSizeConstraintSet

See: updateSizeConstraintSet smart constructor.

Request Lenses

uscsSizeConstraintSetId :: Lens' UpdateSizeConstraintSet Text

The SizeConstraintSetId of the SizeConstraintSet that you want to update. SizeConstraintSetId is returned by CreateSizeConstraintSet and by ListSizeConstraintSets .

uscsChangeToken :: Lens' UpdateSizeConstraintSet Text

The value returned by the most recent call to GetChangeToken .

uscsUpdates :: Lens' UpdateSizeConstraintSet [SizeConstraintSetUpdate]

An array of SizeConstraintSetUpdate objects that you want to insert into or delete from a SizeConstraintSet . For more information, see the applicable data types: * SizeConstraintSetUpdate : Contains Action and SizeConstraint * SizeConstraint : Contains FieldToMatch , TextTransformation , ComparisonOperator , and Size * FieldToMatch : Contains Data and Type

Destructuring the Response

updateSizeConstraintSetResponse

Arguments

:: Int
uscsrsResponseStatus

-> UpdateSizeConstraintSetResponse

Creates a value of UpdateSizeConstraintSetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uscsrsChangeToken - The ChangeToken that you used to submit the UpdateSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

uscsrsResponseStatus - -- | The response status code.

data UpdateSizeConstraintSetResponse

See: updateSizeConstraintSetResponse smart constructor.

Response Lenses

uscsrsChangeToken :: Lens' UpdateSizeConstraintSetResponse (Maybe Text)

The ChangeToken that you used to submit the UpdateSizeConstraintSet request. You can also use this value to query the status of the request. For more information, see GetChangeTokenStatus .

uscsrsResponseStatus :: Lens' UpdateSizeConstraintSetResponse Int

- | The response status code.